Reorganized node and resolver tests
