+++ Forge.Tests/Helpers/StatescriptTestHelpers.cs
@@ -1,7 +1,12 @@
 // Copyright © Gamesmiths Guild.
 #pragma warning disable SA1649, SA1402 // File name should match first type name
 
+using Gamesmiths.Forge.Abilities;
 using Gamesmiths.Forge.Core;
+using Gamesmiths.Forge.Effects;
+using Gamesmiths.Forge.Effects.Components;
+using Gamesmiths.Forge.Effects.Duration;
+using Gamesmiths.Forge.Effects.Magnitudes;
 using Gamesmiths.Forge.Statescript;
 using Gamesmiths.Forge.Statescript.Nodes;
 using Gamesmiths.Forge.Statescript.Nodes.Action;
@@ -39,6 +44,58 @@ internal static class NodeBindings
 		node.BindInput(SetVariableNode.SourceInput, sourcePropertyName);
 		node.BindOutput(SetVariableNode.TargetOutput, targetVariableName, scope);
 		return node;
+	}
+}
+
+/// <summary>
+/// Convenience helpers for creating ability-driven graph contexts in resolver tests.
+/// </summary>
+internal static class ResolverTestContextFactory
+{
+	public static GraphContext CreateAbilityGraphContext(TestEntity entity, float magnitude = 0f)
+	{
+		var graph = new Graph();
+		var captureNode = new CaptureGraphContextNode();
+
+		graph.AddNode(captureNode);
+		graph.AddConnection(new Connection(
+			graph.EntryNode.OutputPorts[EntryNode.OutputPort],
+			captureNode.InputPorts[ActionNode.InputPort]));
+
+		var behavior = new GraphAbilityBehavior(graph);
+
+		AbilityData abilityData = CreateAbilityData("ResolverTest", () => behavior);
+		AbilityHandle? handle = Grant(entity, abilityData);
+		handle!.Activate(out _, magnitude: magnitude);
+
+		return captureNode.CapturedGraphContext!;
+	}
+
+	private static AbilityHandle? Grant(TestEntity target, AbilityData data)
+	{
+		var grantConfig = new GrantAbilityConfig(
+			data,
+			new ScalableInt(1),
+			AbilityDeactivationPolicy.CancelImmediately,
+			AbilityDeactivationPolicy.CancelImmediately,
+			false,
+			false,
+			LevelComparison.Higher);
+
+		var effectData = new EffectData(
+			"Grant",
+			new DurationData(DurationType.Infinite),
+			effectComponents: [new GrantAbilityEffectComponent([grantConfig])]);
+
+		var grantEffect = new Effect(effectData, new EffectOwnership(null, null));
+		_ = target.EffectsManager.ApplyEffect(grantEffect);
+		target.Abilities.TryGetAbility(data, out AbilityHandle? handle);
+		return handle;
+	}
+
+	private static AbilityData CreateAbilityData(string name, Func<IAbilityBehavior> behaviorFactory)
+	{
+		return new AbilityData(name, behaviorFactory: behaviorFactory);
 	}
 }
