+++ Forge.Tests/Statescript/Nodes/State/TimerNodeTests.cs
@@ -8,9 +8,9 @@ using Gamesmiths.Forge.Tests.Helpers;
 
 using static Gamesmiths.Forge.Tests.Helpers.NodeBindings;
 
-namespace Gamesmiths.Forge.Tests.Statescript;
+namespace Gamesmiths.Forge.Tests.Statescript.Nodes.State;
 
-public class StateNodeTests
+public class TimerNodeTests
 {
 	[Fact]
 	[Trait("Graph", "Timer")]
